Extracted PinnedLabel to its own file
ref https://github.com/TryGhost/Ghost/pull/27653

Mirrors how the threading work split out threaded-replies, reply-tree,
and focused-thread into dedicated files. No behaviour change -- the
component is identical, just lifted out of comment.tsx so comment.tsx
doesn't carry the PinnedLabel render logic, its icon imports, or its
labs check inline.
